Instauration routes
Plupart des méthodes placées dans le folder "routes" et dans leur section dédiée en fonction du type de fonctionnalités.
Méthode en cours non déplacées.
Lien avec app.js ok
