+++ views/adminRef/Article.ejs
@@ -13,7 +13,8 @@
                 <form action="/referentielArticle/<%= article._id %>?_method=PUT" method='POST'>
                     <label>Référence: <%= article.reference %></label>
                     <label>Désignation*</label>
-                    <input type='text' name='designation' value="<%= article.designation %>" pattern='[A-Z]{1}[a-z]{1,}' required>
+                    <input type='text' name='designation' value="<%= article.designation %>" required>
+                    <!-- pattern='[A-Z]{1}[a-z]{1,}' -->
                     <label>Prix de vente (€)*</label>
                     <input type='number' name='prix' min='0.00' step='0.01' value="<%= article.prix %>" required>
                     <label>Description</label>
